+++ diag.go
@@ -3,6 +3,7 @@ package main
 import (
 	"bytes"
 	"flag"
+	"fmt"
 	"io/ioutil"
 	"log"
 	"path/filepath"
@@ -84,10 +85,31 @@ func diagLOOPTST() {
 		return
 	}
 
-	log.Printf("Entering LOOPBACK mode - any byte recieved from the Synergy is echo'd back\n")
-	log.Printf("Start the test by pressing RESTORE + RESTORE + PROGRAM 1 on the Synergy\n")
+	fmt.Printf("\nEntering LOOPBACK mode - any byte recieved from the Synergy is echo'd back\n")
+	fmt.Printf("Start the test by pressing RESTORE + RESTORE + PROGRAM 1 on the Synergy\n")
 
 	if err := synio.DiagLOOPTST(); err != nil {
+		log.Printf("ERROR: %s\n", err)
+	}
+}
+
+func diagLINKTST() {
+	flag.Parse()
+
+	log.Printf("%s at %d baud\n",
+		prefsUserPreferences.SerialPort, prefsUserPreferences.SerialBaud)
+
+	if err := synio.Init(prefsUserPreferences.SerialPort, prefsUserPreferences.SerialBaud, true, *serialVerboseFlag); err != nil {
+		log.Printf("ERROR: %s\n", err)
+		return
+	}
+
+	fmt.Printf("\nEntering LINK TEST mode - any byte you type is sent to the Synergy and is \n")
+	fmt.Printf("echo'd back.  LED's on the Synergy show the bytes recieved and various status\n")
+	fmt.Printf("registers of the Synergy's serial connection.\n")
+	fmt.Printf("\nStart the test by pressing RESTORE + PROGRAM 4 on the Synergy\n")
+
+	if err := synio.DiagLINKTST(); err != nil {
 		log.Printf("ERROR: %s\n", err)
 	}
 }