+++ synio/synio_diag.go
@@ -1,8 +1,12 @@
 package synio
 
 import (
-	"log"
+	"bufio"
+	"fmt"
 	"github.com/pkg/errors"
+	"golang.org/x/crypto/ssh/terminal"
+	"log"
+	"os"
 )
 
 func DiagCOMTST() (err error) {
@@ -48,4 +52,43 @@ func DiagLOOPTST() (err error) {
 			return errors.Wrapf(err, "failed to write byte %02x", b)
 		}
 	}
+}
+
+func DiagLINKTST() (err error) {
+
+	state, err := terminal.MakeRaw(int(os.Stdin.Fd()))
+	if err != nil {
+		log.Fatalln("setting stdin to raw:", err)
+		return
+	}
+	defer func() {
+		fmt.Println("Exiting...\n\r")
+		if err := terminal.Restore(int(os.Stdin.Fd()), state); err != nil {
+			log.Println("warning, failed to restore terminal:", err)
+			return
+		}
+	}()
+
+	in := bufio.NewReader(os.Stdin)
+	for {
+		r, _, err := in.ReadRune()
+		if err != nil {
+			log.Println("stdin:", err)
+			break
+		}
+		if r == '\x03' {
+			break
+		}
+		var b = byte(r)
+		if err = serialWriteByte(TIMEOUT_MS, b, "write test byte"); err != nil {
+			log.Fatalln(err, "failed to write byte %02x", b)
+			break
+		}
+		if b, err = serialReadByte(1000*60*5, "read test byte"); err != nil {
+			log.Fatalln(err, "failed to read byte %02x", b)
+			break
+		}
+		fmt.Printf(" sent '%q' (0x%02x) ... recieved 0x%02x (control-C to quit)\n\r", r, r, b)
+	}
+	return
 }